Improve time format shown on web page
Used to be "0d 0h 0m 0s", is now "__:__:00s", "__:01:00s",
"01:00:00s", "1d 00:00:00s".

The idea is that only significant digits should be visible, while
retaining fixed-width time stamps.

The exception to the fixed width is when queue time has been >= 1d,
but that should be uncommon enough not to matter.